find peaks to route, issue with measure menu
